+++ calorie_go_flutter/lib/components/dialog.dart
@@ -1,0 +1,24 @@
+import 'package:flutter/material.dart';
+
+class SampleContent extends StatelessWidget {
+  const SampleContent({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return AlertDialog(
+      backgroundColor: Colors.white,
+      surfaceTintColor: Colors.transparent,
+      content: Column(
+        mainAxisSize: MainAxisSize.min,
+        children: [
+          const Text('Sample Content'),
+          const SizedBox(height: 12),
+          ElevatedButton(
+            onPressed: () => Navigator.of(context).pop(),
+            child: const Text('Close'),
+          ),
+        ],
+      ),
+    );
+  }
+}
