+++ calorie_go_flutter/lib/pages/top/top_page.dart
@@ -3,22 +3,10 @@ import 'package:go_router/go_router.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:serverpod_auth_google_flutter/serverpod_auth_google_flutter.dart';
 
+import '../../components/dialog.dart';
 import '../../constants.dart';
 import '../../gen/assets.gen.dart';
 import 'package:google_fonts/google_fonts.dart';
-
-void main() => runApp(const MyApp());
-
-class MyApp extends StatelessWidget {
-  const MyApp({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return MaterialApp.router(
-      routerConfig: routerConfig,
-    );
-  }
-}
 
 class TopPage extends HookConsumerWidget {
   const TopPage({super.key});
@@ -49,20 +37,22 @@ class TopPage extends HookConsumerWidget {
                 redirectUri: Uri.parse('http://localhost:8082/googlesignin'),
                 onFailure: () =>
                     throw Exception('Failed to sign in with Google.'),
-                onSignedIn: () => _showDialog(context),
+                onSignedIn: () async {
+                  await showDialog(
+                    context: context,
+                    builder: (BuildContext context) {
+                      return const SampleContent();
+                    },
+                  );
+                  if (!context.mounted) return;
+                  context.go('/');
+                },
               ),
               const SizedBox(height: 70),
             ],
           ),
         ),
       ),
-    );
-  }
-
-  void _showDialog(BuildContext context) {
-    showDialog(
-      context: context,
-      builder: (_) => const SampleContent(),
     );
   }
 }
@@ -90,48 +80,6 @@ class _Header extends StatelessWidget {
           ),
         ),
       ],
-    );
-  }
-}
-
-final routerConfig = GoRouter(
-  routes: [
-    GoRoute(
-      path: RoutingConfig.home.path,
-      name: RoutingConfig.home.name,
-      pageBuilder: (_, __) => const MaterialPage(child: TopPage()),
-    ),
-  ],
-);
-
-enum RoutingConfig {
-  home('/', 'home');
-
-  final String path;
-  final String name;
-
-  const RoutingConfig(this.path, this.name);
-}
-
-class SampleContent extends StatelessWidget {
-  const SampleContent({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return AlertDialog(
-      backgroundColor: Colors.white,
-      surfaceTintColor: Colors.transparent,
-      content: Column(
-        mainAxisSize: MainAxisSize.min,
-        children: [
-          const Text('Sample Content'),
-          const SizedBox(height: 12),
-          ElevatedButton(
-            onPressed: () => Navigator.of(context).pop(),
-            child: const Text('Close'),
-          ),
-        ],
-      ),
     );
   }
 }
